requirements.txt
wyoming==1.2.0
azure-cognitiveservices-speech==1.32.1